topbadan

+++ Manager_Bot/main.py
@@ -1,5 +1,6 @@
 import os
 import shutil
+from sys import argv
 from time import sleep
 
 
@@ -19,7 +20,7 @@ def exit():
 def onExit():
     dirs = os.listdir()
     for i in dirs:
-        if not '.py' in i and i != 'assets':
+        if not '.py' in i and i != 'assets' and i != 'heroku':
             while True:
                 try:
                     shutil.rmtree(i)
@@ -29,11 +30,22 @@ def onExit():
     print('onExit')
 
 
+def onBoot():
+    sleep(0.1)
+
+
 if __name__ == "__main__":
-    start()
-    try:
-        while True:
-            loop()
-            sleep(0.02)
-    except:
-        onExit()
+    argv.pop(0)
+    if argv[0] == 'start':
+        start()
+        try:
+            while True:
+                loop()
+                sleep(0.02)
+        except:
+            onExit()
+    elif argv[0] == 'onboot':
+        onBoot()
+    else:
+        print('Error: There is no argument named ' + str(argv[0]))
+    sleep(1)
